Assignments 1-4 completed
Assignments are done and changes are commited and pushed.
I used buider pattern for Product.java and ProductModel.java.
The builder pattern simplifies the creation of objects and also simplifies the code as you dont have to call a complex constructor or call several setter methods
on the created object.
It can be applied to all other classes from package model, but it's not necessary for this task.
Some of getters and setters are not used in this specific project, but I didn't remove it
(it's better to have it, for example, in case of changing some parts of code in project that requires that getter or setter)

For Assignment 4 I created two methods for extracting that new described data:
1. extractAndModelNewInformationFromProduct()
For this first method, new class is created to represent extracted informations - ProductModel.java
2. extractNewInformationsFromProduct()
This second method uses simple presentation of extracted data.
Currently, call of this method is placed in MSApplication contructor that is called by main method and they are commented by "//".
In order to run them, just remove comment sign one by one and run main method and you will see small difference in presentation of data
, but in the code, for one call class ProductModel is used, and for other simple String presentation.